Cucumber failing on status feature
I normally will not commit and push a failing spec or
feature, but in this case, the status.feature has likely
been borked for quite a while. Not sure it's worth saving,
or it may be something better handled with a unit test.
Probably ought to break out the Cucumber book on this one.