rails502 - need to fix controller permitted params
This commit is a stopping point, the next step is to
get the parameter permitting unscrewed in the UserController,
and that's work for another day. I'm writing this here in
the commit message as it saves writing it in a todo list
on paper, which migh get lost.

TODO: user controller permitted params