major derp moment here with response.body
I have waling on this stupid <title> selector check
in several rails projects, and have been stymied across
the board, until now.

As part of my TDD practice, in the past I would specify
what the <title> elements should be, then ensure that they
were implemented appropriately. After everything is done,
these tests and specs for <title> elements admittedly look
a lot like "framework" tests, but that has never been the
intention.

It turns out, the `response` object could be processed
such the the `response.body` was automatically checked
for the relevant html, which was then duly processed by
capybara. So, all of these specs worked in Rails 3.*, but
stopped working in Rails 4.*. At the time, I didn't feel
all the committed to any of the projects, having my hands
full with work commitments. Now with the Rails 5 upgrade
underway, I am very glad I set those specs to pending in-
stead of deleting them.

Whether or not testing for the value of <title> is ultimately
a good idea, updating the specs to work has been worthwhile
to me.