user index view spec fix
Sort of a fix to make it sort of work. Granted the whole
notion of view spec is sketchy, but since the specs are
already in place, and rspec continues to support them, it's
worth it to me to keep up with the changes in thinking with
respect to view specs as this thinking evolves over time.

The above is almost certainly post-facto justification, but
whatever.

There is a deeper lesson here, which revolves around the
decisions on how view spec implementation was designed. I
expect to get more into that when I go back through and
update all the view specs in my applications, a task I'll
start once all the pending specs are updated, Rails 5 is
running, etc.