small cleanups and gems
First, cannot upgrade to rails 5.1 as cucumber-rails has
a dependency on railties < 5.1. This is a major bummer, but
not a show stopper. I'm not really using cucumber in anger
at the moment anyway.

A few other gems such as jasmine and haml-rails were updated.
Jasmine will need to be watched, as it depends on phantomjs,
which is no longer under development by its main programmer.
Not sure if anyone else has picked that up yet.

Status feature spec continues to mystify, will need to fire
up the application locally to investigate.