envs-51 - application and assets updated
The app:update is starting down the homestretch now.
Specs and development server run (apparently) correctly.

This is probably good enough to merge back into master.

Should consider squashing some of these commits, but the
commit order is actually pretty useful for helping update
the other Rails applications I have going on at the moment.
So I may leave it as-is.